code refactor
지금 메인 씬에서 카메라를 생성해서 렌더 컴포넌트에 카메라를 넘겨주고 있어 그리고 추가고 플레이어를 개발해서 이동로직을 추가하려하는데 플레이어 게임오브젝트의 업데이트에서 하려고 했어 근데 생각해보니 컴포넌트 기능이 있기 때문에 플레이어 컴포넌트를 만들어서 거기 업데이트에 이동로직을 추가하려고 수정했어 그러면 플레이어 게임오브젝트에서는 가지고 있는 컴포넌트의 업데이트들을 전부 실행하면 돼 그리고 그럴거면 렌더또한 마찬가지야 그럼 렌더에서는 카메라가 필요한데 카메라는 메인씬에 있어 이럴때 카메라를 어떻게 하지

정리하자면 기존
메인 씬 > 게임오브젝트들 순회하며 렌더 진행 파라미터 카메라
수정
메인 씬 > 게임오브젝트 순회 렌더 > 컴포넌트 순회 렌더 (카메라 파라미터 필요)

카메라를 파라미터로 다 넘겨주게 변경

+++ scene/lobby_scene.cpp
@@ -10,7 +10,7 @@ void LobbyScene::Init() {
     Ui_customize::Initialization();
 }
 
-void LobbyScene::Update(double deltaTime) {
+void LobbyScene::Update(float deltaTime) {
     if(Input::GetKey(eKeyCode::A)) {
         SPDLOG_INFO("A 눌림");
     }
