+++ scene/main_scene.h
@@ -11,7 +11,7 @@
 class MainScene : public Scene {
 public:
     void Init() override;
-    void Update(double deltaTime) override;
+    void Update(float deltaTime) override;
     void Render() override;
     void Shutdown() override;
